0.16.08 - WHY WASS THAT SO EASY ;-; - rounded x positions of right portion of divided entities to prevent sub-pixel imprecision - collider for text area now updates upon text change

+++ src/dividedentity.cpp
@@ -2,7 +2,7 @@
 
 void DividedEntity::SetPos(Vector2f& p_pos){
     top -> setPos(Vector2f(p_pos.x + left -> getCurrentFrame().w * left -> getScale().x, p_pos.y));
-    topRight -> setPos(top -> getPos() + Vector2f(top -> getCurrentFrame().w * top -> getScale().x, 0));
+    topRight -> setPos(top -> getPos() + Vector2f((int)(top -> getCurrentFrame().w * top -> getScale().x), 0));
     left -> setPos(p_pos + Vector2f(0, left -> getCurrentFrame().y * top -> getScale().y));
     center -> setPos(Vector2f(top -> getPos().x, left -> getPos().y));
     right -> setPos(Vector2f(topRight -> getPos().x, left -> getPos().y));
